pyproject.toml
[tool.poetry]
name = "slo-lr-detection"
version = "1.0.0"
description = "Setting up pipeline for training on GCP"
authors = ["Javad Hamidi Esfahani"]
readme = "README.md"
packages = [
    { include = "slo-lr-detection", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.12,<3.13.0"
toml = "^0.10.2"
numpy = "^1.26.4"
opencv-python = "^4.9.0.80"
torch = "^2.5.0"
torchsummary = "^1.5.1"
google = "^3.0.0"
kfp = "^2.9.0"
ml_collections = "^0.1.1"
torchvision = "^0.20.1"
matplotlib = "^3.9.2"
tqdm = "^4.66.6"
mlflow = "^2.17.2"
cloudml-hypertune = "^0.1.0.dev6"
google-cloud-aiplatform = "^1.72.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.dev-dependencies]
pytest = "^7.0"
pylint = "^3.3.1"
